+++ gazebo_plugins/include/gazebo_plugins/gazebo_ros_diff_drive.h
@@ -81,6 +81,7 @@ namespace gazebo {
       GazeboRosDiffDrive();
       ~GazeboRosDiffDrive();
       void Load(physics::ModelPtr _parent, sdf::ElementPtr _sdf);
+      void Reset();
 
     protected:
       virtual void UpdateChild();